requirements.txt
Flask
requests
python-dotenv
Pillow
numpy